+++ internal/modules/order/create/dto.go
@@ -3,7 +3,7 @@ package create
 import "PocketArtisan/internal/entities"
 
 type NewOrderRequest struct {
-	CraftsmanID     uint                  `json:"craftsman_id" binding:"required"`
+	CraftsmanID     uint64                  `json:"craftsman_id" binding:"required"`
 	Items           []NewOrderItemRequest `json:"items" binding:"required,min=1,dive,required"`
 	PaymentType     entities.PaymentType  `json:"payment_type" binding:"required"`
 	ShippingAddress string                `json:"shipping_address" binding:"required"`
@@ -13,7 +13,7 @@ type NewOrderRequest struct {
 //additional data for customer will be fetched from the database using the customer ID, so we don't need to include it in the request. The same goes for craftsman data.
 
 type NewOrderItemRequest struct {
-	ProductID uint `json:"product_id" binding:"required"`
+	ProductID uint64 `json:"product_id" binding:"required"`
 	Quantity  int  `json:"quantity" binding:"required,gt=0"`
 }
